+++ src/utils/filenamehandler.cpp
@@ -41,7 +41,7 @@ QString FileNameHandler::parseFilename(const QString &name) {
         char data[MAX_CHARACTERS] = {0};
         std::strftime(data, sizeof(data),
                       tempData, std::localtime(&t));
-        res = QString::fromLocal8Bit(data, strlen(data));
+        res = QString::fromLocal8Bit(data, (int)strlen(data));
         free(tempData);
     }
     // add the parsed pattern in a correct format for the filesystem
